Added more send configuration benchmarks
