+++ .github/workflows/build-windows.yml
@@ -273,7 +273,7 @@ jobs:
       fail-fast: false
       matrix:
         # List of all benchmark types
-        bm_type: [pubsub, pubsub_multi, service, setup]
+        bm_type: [pubsub, pubsub_config, pubsub_multi, service, setup]
     
     env:
       # Name of the benchmark log file without extension
@@ -318,7 +318,7 @@ jobs:
         path: ${{ github.workspace }}/ecal_benchmark_${{ matrix.bm_type }}_log.json
     
     - name: Calculate Throughput
-      if: ${{ matrix.bm_type == 'pubsub' || matrix.bm_type == 'pubsub_multi' }}
+      if: startsWith(matrix.bm_type, 'pubsub')
       run: |
         python throughput_calculator.py -f ${{ github.workspace }}/${{ env.LOG_FILE_NAME }}.json
       working-directory: ${{ github.workspace }}
@@ -337,7 +337,7 @@ jobs:
       shell: pwsh
         
     - name: Log Calculated Throughput with Bencher
-      if: ${{ matrix.bm_type == 'pubsub' || matrix.bm_type == 'pubsub_multi' }}
+      if: startsWith(matrix.bm_type, 'pubsub')
       run: |
         bencher run `
         --project ecal `
