+++ ecal/benchmarks/pubsub_config/CMakeLists.txt
@@ -18,10 +18,18 @@
 
 cmake_minimum_required(VERSION 3.15)
 
-find_package(benchmark REQUIRED)
+project(ecal_benchmark_pubsub_config)
 
-add_subdirectory(pubsub)
-add_subdirectory(pubsub_config)
-add_subdirectory(pubsub_multi)
-add_subdirectory(service)
-add_subdirectory(setup)
+set(source_files
+  benchmark_pubsub_config.cpp
+)
+
+add_executable(${PROJECT_NAME} ${source_files})
+
+target_link_libraries(${PROJECT_NAME}
+  PRIVATE
+    eCAL::core
+    benchmark::benchmark
+)
+
+target_compile_features(${PROJECT_NAME} PRIVATE cxx_std_14)
